Moved concurrent test to AbstractTestStore
